Move signal_level to model package
