+++ example/lib/model/signal_level.dart
